requirements.txt
# Core LLM & AI
openai>=1.0.0
anthropic>=0.8.0

# CLI & Interface
rich>=13.0.0
click>=8.0.0
prompt-toolkit>=3.0.0
typer>=0.9.0

# Web Framework - Modern UI
fastapi>=0.100.0
uvicorn>=0.20.0
websockets>=11.0.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Tools & Integrations
requests>=2.31.0
beautifulsoup4>=4.12.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# File Operations
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
aiofiles>=23.0.0

# Database
sqlalchemy>=2.0.0

# Async & Concurrency
aiohttp>=3.8.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Monitoring & Logging
structlog>=23.0.0
prometheus-client>=0.17.0

# Security
cryptography>=41.0.0